pyproject.toml
[project]
name = "confantic"
version = "1.0.0"
description = "A textual editor and validator for YAML/JSON files against Pydantic models."
authors = [{ name = "Kevin Welsh", email = "kevin.welsh@verdantevolution.com" }]
readme = "README.md"
requires-python = ">=3.8"
dependencies = ["pydantic", "pyyaml", "textual[syntax]>=0.73.0"]

[project.scripts]
confantic = "confantic.__main__:main"
